feat(client): fix update bugs refactor and remove unused functions types etc
